Using latest snapshot of junit5
Moves everything to latest snapshot, which cleaned up some of the code
and formatting. Moved a bunch of common code to Clojure in a support
project (lang-clojure) that could get reused for other test frameworks.

Still want to convert this to a HierarchicalTestEngine and use the Node
abstraction. Should consider how to handle dynamic tests as well.