Ensure fixtures run appropriately
As noted in a leiningen issue[1], overriding test-var isn't an
appropriate filtering approach, as it doesn't stop fixtures from
running for the filtered-out tests. They pointed to a Clojure
issue[2] that added the test-vars fn, which runs the fixtures
only for the vars you pass in. However, they indicated later in
that thread that while they waited for that issue to be resolved
(which it is now), they would just remove the :test metadata from
the vars. I haven't looked to confirm which approach they took,
so not sure if I'm being consistent. The only drawback seems to
be the lack of support fo the :test-ns-hook feature that allows
you to control order. This doesn't seem like that great of a
feature anyway, so I'm not too bothered.

[1] https://github.com/technomancy/leiningen/issues/1269
[2] http://dev.clojure.org/jira/browse/CLJ-866