Really hacky executer working-ish
Needed to send started events for the suites instead of just the
base tests. Need to clean up the logging quite a bit and also have
it report the correct result of the tests. May want to just skip
to writing my own launcher wrapper now to pass the information back
in a more helpful way.